+++ src/poset/badger_store_test.go
@@ -426,7 +426,7 @@ func TestDBFrameMethods(t *testing.T) {
 		event.Sign(p.privKey)
 		events[id] = event.Message
 
-		root := NewBaseRoot(uint32(id))
+		root := NewBaseRoot(uint64(id))
 		roots[p.hex] = root
 	}
 	frame := &Frame{
@@ -532,7 +532,7 @@ func TestBadgerEvents(t *testing.T) {
 		}
 	}
 
-	expectedKnown := make(map[uint32]int64)
+	expectedKnown := make(map[uint64]int64)
 	for _, p := range participants {
 		expectedKnown[p.id] = testSize - 1
 	}
@@ -690,7 +690,7 @@ func TestBadgerFrames(t *testing.T) {
 		event.Sign(p.privKey)
 		events[id] = event.Message
 
-		root := NewBaseRoot(uint32(id))
+		root := NewBaseRoot(uint64(id))
 		roots[p.hex] = root
 	}
 	frame := &Frame{
