+++ src/poset/root.go
@@ -64,7 +64,7 @@ ex 2:
 
 // NewBaseRootEvent creates a RootEvent corresponding to the the very beginning
 // of a Poset.
-func NewBaseRootEvent(creatorID uint32) RootEvent {
+func NewBaseRootEvent(creatorID uint64) RootEvent {
 	hash := fmt.Sprintf("Root%d", creatorID)
 	res := RootEvent{
 		Hash:             hash,
@@ -93,7 +93,7 @@ func (re *RootEvent) Equals(that *RootEvent) bool {
 // Root's Others.
 
 // NewBaseRoot initializes a Root object for a fresh Poset.
-func NewBaseRoot(creatorID uint32) *Root {
+func NewBaseRoot(creatorID uint64) *Root {
 	rootEvent := NewBaseRootEvent(creatorID)
 	res := &Root{
 		NextRound:  0,
